[Feat] Add ai start time

+++ app/src/main/java/com/fitsionary/momspt/data/api/response/VideoCheckTime.kt
@@ -6,6 +6,6 @@ import android.os.Parcelable
 
 @Parcelize
 data class VideoCheckTime(
-    val workoutFinishTime: Int,
-    val workoutStartTime: Int
+    val workoutFinishTime: Long,
+    val workoutStartTime: Long
 ) : Parcelable
